requirements.txt
click==8.1.3
Flask==2.2.2
graphyte==1.7.1
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
waitress==2.1.2
Werkzeug==2.2.2
zipp==3.8.1
